Move locker to the `model` namespace
