+++ phibes/model/__init__.py
@@ -1,5 +1,17 @@
 from . item import FILE_EXT, Item
+from . locker import Locker, LOCKER_FILE
 from . schema import Schema
 from . secret import Secret
 from . tag import Tag
 from . template import Template
+
+__all__ = [
+    "FILE_EXT",
+    "Item",
+    "Locker",
+    "LOCKER_FILE",
+    "Schema",
+    "Secret",
+    "Tag",
+    "Template"
+]
