+++ tests/lib/locker_helper.py
@@ -11,8 +11,7 @@ from phibes.lib.config import CONFIG_FILE_NAME
 from phibes.lib.config import ConfigModel, set_home_dir
 from phibes.lib.config import load_config_file, write_config_file
 from phibes.lib.errors import PhibesNotFoundError
-from phibes.lib.locker import Locker
-from phibes.model import Item
+from phibes.model import Item, Locker
 
 
 class BaseTestClass(object):
